third_party/blink/web_tests/fast/events/init-event-exceptions.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description("Tests the type checking of init*Event() methods.");
shouldThrow("document.createEvent('CompositionEvent').initCompositionEvent('', false, false, {}, '')");
shouldThrow("document.createEvent('KeyboardEvent').initKeyboardEvent('', false, false, {}, '', 0, false, false, false, false)");
shouldThrow("document.createEvent('MouseEvent').initMouseEvent('', false, false, null, 0, 0, 0, 0, 0, false, false, false, false, 0, {})");
shouldThrow("document.createEvent('MouseEvent').initMouseEvent('', false, false, {}, 0, 0, 0, 0, 0, false, false, false, false, 0, null)");
shouldThrow("document.createEvent('TextEvent').initTextEvent('', false, false, {}, '')");
shouldThrow("document.createEvent('UIEvent').initUIEvent('', false, false, {}, 0)");
</script>
